// budget-api/src/notifications/entities/notification-type.enum.ts
export enum NotificationType {
  BILL_OVERDUE = 'BILL_OVERDUE',
  BILL_DUE_SOON = 'BILL_DUE_SOON',
  BUDGET_THRESHOLD = 'BUDGET_THRESHOLD',
  BUDGET_EXCEEDED = 'BUDGET_EXCEEDED',
  LOW_BALANCE = 'LOW_BALANCE',
  LARGE_TRANSACTION = 'LARGE_TRANSACTION',
  RECURRING_BILL_ENDING = 'RECURRING_BILL_ENDING',
  WORKSPACE_INVITATION = 'WORKSPACE_INVITATION',
  MONTHLY_SUMMARY = 'MONTHLY_SUMMARY',
}